+++ packages/transaction-pay-controller/src/strategy/relay/solana-pay.ts
@@ -337,7 +337,15 @@ export function mapRelayStatus(
     return 'refund';
   }
 
-  return 'pending';
+  if (
+    ['waiting', 'depositing', 'pending', 'submitted', 'delayed'].includes(
+      status,
+    )
+  ) {
+    return 'pending';
+  }
+
+  return 'unknown';
 }
 
 function getSourceChainId(source: TransactionPaySource): CaipChainId {